contrib/MarranoFTP: Minor delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@43444 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/Networking/Apps/MarranoFTP/main.c
@@ -3254,7 +3254,7 @@ BOOL RefreshLocalView(Connection *cn)
                     if (fib->fib_DirEntryType == ST_USERDIR) {
                         vc.size = BufferAddStr(&lv->ListBuffer, "<DIR>");
                     } else {
-                        snprintf(temp, 511, "%ld", fib->fib_Size);
+                        snprintf(temp, 511, "%ld", (long)fib->fib_Size);
                         vc.size = BufferAddStr(&lv->ListBuffer, temp);
                     }
